Move TradeSide and TransferType types to contract-exchange package
