+++ contracts/exchange/test/utils/fill_order_simulator.ts
@@ -2,8 +2,6 @@ import { constants, orderUtils } from '@0x/contracts-test-utils';
 import {
     Order,
     SupportedProvider,
-    TradeSide,
-    TransferType,
 } from '@0x/order-utils';
 import { FillResults } from '@0x/types';
 import { BigNumber } from '@0x/utils';
@@ -13,6 +11,10 @@ import {
     AbstractBalanceAndProxyAllowanceLazyStore as LazyStore,
 } from './abstract/abstract_balance_and_proxy_allowance_lazy_store';
 import { ExchangeTransferSimulator } from './exchange_transfer_simulator';
+import {
+    TradeSide,
+    TransferType,
+} from './types';
 
 export enum FillOrderError {
     OrderUnfillable = 'ORDER_UNFILLABLE',
