+++ packages/order-utils/src/types.ts
@@ -5,16 +5,6 @@ export enum TypedDataError {
     InvalidMetamaskSigner = "MetaMask provider must be wrapped in a MetamaskSubprovider (from the '@0x/subproviders' package) in order to work with this method.",
 }
 
-export enum TradeSide {
-    Maker = 'maker',
-    Taker = 'taker',
-}
-
-export enum TransferType {
-    Trade = 'trade',
-    Fee = 'fee',
-}
-
 export interface CreateOrderOpts {
     takerAddress?: string;
     senderAddress?: string;
